requirements.txt
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.8.0
xgboost==2.0.3
pandas==2.1.4
numpy==1.26.2
yfinance==0.2.33
streamlit==1.29.0
python-dotenv==1.0.0
scikit-learn==1.3.2
plotly==5.18.0
sentence-transformers==2.2.2
ollama==0.1.6

# Added for OpenAI embeddings, model access and utilities
openai>=0.27.0
langchain-openai==0.0.6
tiktoken>=0.4.0
joblib>=1.3.2
tqdm>=4.66.1
requests>=2.31.0
chromadb>=0.4.22
robin-stocks>=3.0.0
